Add ut to check problem

+++ transaction-executor/bcos-transaction-executor/precompiled/PrecompiledImpl.h
@@ -96,7 +96,7 @@ inline EVMCResult callPrecompiled(Precompiled const& precompiled, auto& storage,
                 auto buffer = std::unique_ptr<uint8_t>(new uint8_t[params->m_execResult.size()]);
                 std::copy(params->m_execResult.begin(), params->m_execResult.end(), buffer.get());
                 EVMCResult result{evmc_result{
-                    .status_code = (evmc_status_code)(int32_t)protocol::TransactionStatus::None,
+                    .status_code = (evmc_status_code)protocol::TransactionStatus::None,
                     .gas_left = response->m_gasLeft,
                     .gas_refund = 0,
                     .output_data = buffer.release(),
